agents/plumber-data-engineering-assistant/pyproject.toml
[project]
name = "plumber"
version = "0.1.0"
description = "A data engineering agent buddy, which help deploy, monitor and make custom data pipelines for developers powered by Gemini Models and ADK Deploy Dataproc (Spark), Dataflow (Beam) and dBT Pipelines directly with human-in-the-loop (list) Interactively check logs , monitor metrics and status of pipeline; no need to go through multiple tabs and screens Use Github tool interactively to commit code directly to your repository Add custom transformation rule by just using simple source to target mapping sheet for Dataproc and dBT jobs."
authors = [
    { name = "Anish Sarangi", email = "anishks@google.com" },
    { name = "Mayank Kalsan", email = "mayankkalsan@google.com" },
    { name = "Aditya", email = "itsadi@google.com" },
    { name = "Bala Priya", email = "baswalbalapri@google.com" },
    { name = "Aman Usmani", email = "amanusmani@google.com" },
    { name = "Pradeep Kumar K", email = "pradeepkpk@google.com" },
    { name = "Bhabani Mahapatra ", email = "bhabaniranjan@google.com" }
    ]
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.12",
]
readme = "README.md"
requires-python = ">=3.10,<3.13"

dependencies = [
    "beautifulsoup4==4.13.4",
    "dbt-bigquery==1.10.1",
    "fastapi==0.116.1",
    "uvicorn==0.35.0", # For running FastAPI
    "GitPython==3.1.45",
    "google-adk==1.16.0",
    "google-api-python-client==2.177.0",
    "google-cloud-monitoring==2.27.2",
    "google-cloud-dataform==0.7.0",
    "google-cloud-storage==2.19.0",
    "google-generativeai==0.8.5",
    "Pillow==11.3.0",
    "python-dotenv==1.1.1",
    "requests==2.32.4",
    "apache-beam[gcp]==2.69.0",
]

[project.optional-dependencies]
lint = [
    "ruff>=0.4.6",
    "mypy>=1.15.0",
    "codespell>=2.2.0",
]

dev = [
    "plumber[lint]", # Installs all linting dependencies
    "ipykernel>=6.29.5",
    "ipython>=8.36.0",
    "pytest>=8.3.5",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.26.0",
    "pre-commit>=4.2.0",
    "deptry>=0.23.0",
]

[build-system]
requires = [
 "setuptools>=80.9.0",
 "wheel"
]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py312" # Matching your requires-python

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # pyflakes
    "W",  # pycodestyle warnings
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
    "RUF",# ruff specific rules
]
ignore = ["E501", "C901"] # ignore line too long, too complex

[tool.ruff.lint.isort]
known-first-party = ["plumber"] # Set to your project's name

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]
exclude = [".venv"]

[tool.codespell]
ignore-words-list = "rouge"
skip = ".venv,uv.lock"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"
markers = [
    "slow: marks tests as slow",
    "infra_ready: marks tests as requiring cloud infra to be ready",
]

[tool.agent_packs]
template = "adk"